fakenvapi: spoof PCI identifiers and add NvAPI_GPU_GetMemoryInfo (24GB)
007 First Light gates RT/PT options on NVIDIA GPU with >= 12GB VRAM.
nvapi-side identity was inconsistent (name said RTX 4090, PCI ids returned
the real AMD adapter, memory info was unimplemented). Report the spoofed
NVIDIA identity and a 24GB memory budget through nvapi as well.

+++ OptiScaler/NVNGX_Parameter.cpp
@@ -419,7 +419,7 @@ template <typename T> NVSDK_NGX_Result NVNGX_Parameters::getT(const char* key, T
         if (k == m_values.end())
             LOG_INFO("NGX param READ: '{}' -> MISSING", key);
         else
-            LOG_INFO("NGX param READ: '{}' -> int {}", key, (int) (long long) (*k).second);
+            LOG_INFO("NGX param READ: '{}' -> int {}", key, static_cast<int>((*k).second));
     }
 
     if (k == m_values.end())
